RSA Factoring

+++ factor.c
@@ -2,7 +2,7 @@
 
 int main(int argc, char *argv[])
 {
-    unsigned long num;
+    unsigned long long num;
     FILE *file = fopen(argv[1], "r");
 
     if (argc != 2)
@@ -17,7 +17,7 @@ int main(int argc, char *argv[])
         return (1);
     }
 
-    while (fscanf(file, "%lu", &num) != EOF)
+    while (fscanf(file, "%llu", &num) != EOF)
     {
         if (num > 1)
         {
