+++ factorize.c
@@ -1,14 +1,14 @@
 #include "factor.h"
 
-void factorize(unsigned long num)
+void factorize(unsigned long long num)
 {
-    unsigned long i;
+    unsigned long long i;
 
     for(i = 2; i <= num; i++)
     {
         if (num % i == 0)
         {
-            printf ("%lu=%lu*%lu\n", num, i, num / i);
+            printf ("%llu=%llu*%llu\n", num, i, num / i);
             return;
         }
     }
